# docs/swagger.yaml
basePath: /
definitions:
  buildinfo.Info:
    properties:
      build_time:
        type: string
      commit:
        type: string
      started_at:
        type: string
      uptime:
        type: string
      version:
        type: string
    type: object
  catalog.Source:
    properties:
      available:
        type: boolean
      error:
        type: string
      name:
        type: string
      origin:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  health.Dependency:
    properties:
      latency:
        type: string
      status:
        type: string
    type: object
  health.Status:
    properties:
      dependencies:
        additionalProperties:
          $ref: '#/definitions/health.Dependency'
        type: object
      status:
        type: string
    type: object
  httptransport.GetSwaggerSpecRequest:
    properties:
      name:
        example: platform--identity-service
        type: string
    required:
    - name
    type: object
  httptransport.ListSwaggerServicesRequest:
    properties:
      keyword:
        type: string
      page:
        minimum: 1
        type: integer
      page_size:
        maximum: 100
        minimum: 1
        type: integer
    type: object
  httptransport.MeResponseBody:
    properties:
      subject:
        type: string
    type: object
  httptransport.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
      request_id:
        type: string
    type: object
  httptransport.SwaggerServicesBody:
    properties:
      items:
        items:
          $ref: '#/definitions/catalog.Source'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  httptransport.SwaggerSpecBody:
    properties:
      document:
        additionalProperties: {}
        type: object
      name:
        type: string
    type: object
info:
  contact: {}
  description: 'Discovers and aggregates platform OpenAPI documents. Application error
    codes: 0 success; 10000-19999 common/input; 20000-29999 authentication/authorization;
    30000-39999 business; 50000-59999 infrastructure.'
  title: Platform Swagger Service API
  version: "1.0"
paths:
  /api/v1/me:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.MeResponseBody'
              type: object
        "401":
          description: 'Code 20001: unauthorized'
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Return the authenticated subject
      tags:
      - authentication
  /api/v1/swagger/services:
    post:
      consumes:
      - application/json
      parameters:
      - description: Search and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListSwaggerServicesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.SwaggerServicesBody'
              type: object
        "401":
          description: 'Code 20001: unauthorized'
          schema:
            $ref: '#/definitions/httptransport.Response'
        "403":
          description: 'Code 20003: forbidden'
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: List discovered OpenAPI services
      tags:
      - OpenAPI aggregation
  /api/v1/swagger/spec:
    post:
      consumes:
      - application/json
      parameters:
      - description: Service source name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.GetSwaggerSpecRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.SwaggerSpecBody'
              type: object
        "400":
          description: 'Code 10001: invalid request'
          schema:
            $ref: '#/definitions/httptransport.Response'
        "401":
          description: 'Code 20001: unauthorized'
          schema:
            $ref: '#/definitions/httptransport.Response'
        "403":
          description: 'Code 20003: forbidden'
          schema:
            $ref: '#/definitions/httptransport.Response'
        "404":
          description: 'Code 10004: source not found'
          schema:
            $ref: '#/definitions/httptransport.Response'
        "503":
          description: 'Code 50003: source unavailable'
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Get one aggregated OpenAPI document
      tags:
      - OpenAPI aggregation
  /api/v1/version:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/buildinfo.Info'
              type: object
      summary: Return build and runtime version information
      tags:
      - operations
  /live:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/health.Status'
              type: object
      summary: Check process liveness
      tags:
      - operations
  /ready:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/health.Status'
              type: object
        "503":
          description: 'Code 50003: dependency unavailable'
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/health.Status'
              type: object
      summary: Check database and Redis readiness
      tags:
      - operations
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Enter "Bearer {JWT}".
    in: header
    name: Authorization
    type: apiKey
  PSK:
    description: Enter "PSK {shared-key}" for routes configured with PSK authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"